Added: separate product util for fetching identification value(#2y8n6qv)
