+++ src/utils/index.ts
@@ -1,6 +1,7 @@
 import { toastController } from '@ionic/vue';
 import { translate } from '@/i18n'
 import { Plugins } from '@capacitor/core';
+import productHelpers from './product'
 
 // TODO Use separate files for specific utilities
 
@@ -29,25 +30,4 @@ const copyToClipboard = async (text: string) => {
     });
 }
 
-const getProductIdentificationValue = (id: string, product: any) => {
-
-  // handled this case as on page load initially the data is not available, so not to execute furthur code
-  // untill product are available
-  if(!Object.keys(product).length) {
-    return;
-  }
-
-  let value = product[id]
-
-  // considered that the goodIdentification will always have values in the format "id/value" and there will be no entry like "id/"
-  const identification = product['goodIdentifications'].find((identification: string) =>  identification.startsWith(id + "/"))
-
-  if(identification) {
-    const goodIdentification = identification.split('/')
-    value = goodIdentification[1]
-  }
-
-  return value;
-}
-
-export { showToast, hasError, copyToClipboard, getProductIdentificationValue }
+export { showToast, hasError, copyToClipboard, productHelpers }
