+++ src/views/PurchaseOrderDetail.vue
@@ -46,8 +46,8 @@
                   <Image :src="getProduct(item.productId).mainImageUrl" />
                 </ion-thumbnail>
                 <ion-label class="ion-text-wrap">
-                  <h2>{{ getProductIdentificationValue(productIdentificationPref.primaryId, getProduct(item.productId)) }}</h2>
-                  <p>{{ getProductIdentificationValue(productIdentificationPref.secondaryId, getProduct(item.productId)) }}</p>
+                  <h2>{{ productHelpers.getProductIdentificationValue(productIdentificationPref.primaryId, getProduct(item.productId)) }}</h2>
+                  <p>{{ productHelpers.getProductIdentificationValue(productIdentificationPref.secondaryId, getProduct(item.productId)) }}</p>
                 </ion-label>
               </ion-item>
             </div>
@@ -134,7 +134,7 @@ import Scanner from "@/components/Scanner.vue"
 import AddProductToPOModal from '@/views/AddProductToPOModal.vue'
 import LocationPopover from '@/components/LocationPopover.vue'
 import ImageModal from '@/components/ImageModal.vue';
-import { copyToClipboard, getProductIdentificationValue } from '@/utils';
+import { copyToClipboard, productHelpers } from '@/utils';
 
 export default defineComponent({
   name: "PurchaseOrderDetails",
@@ -264,7 +264,7 @@ export default defineComponent({
       checkmarkDone,
       copyOutline,
       copyToClipboard,
-      getProductIdentificationValue,
+      productHelpers,
       router,
       saveOutline,
       store,
