+++ src/views/ReturnDetails.vue
@@ -42,8 +42,8 @@
                   <Image :src="getProduct(item.productId).mainImageUrl" />
                 </ion-thumbnail>
                 <ion-label class="ion-text-wrap">
-                  <h2>{{ getProductIdentificationValue(productIdentificationPref.primaryId, getProduct(item.productId)) }}</h2>
-                  <p>{{ getProductIdentificationValue(productIdentificationPref.secondaryId, getProduct(item.productId)) }}</p>
+                  <h2>{{ productHelpers.getProductIdentificationValue(productIdentificationPref.primaryId, getProduct(item.productId)) }}</h2>
+                  <p>{{ productHelpers.getProductIdentificationValue(productIdentificationPref.secondaryId, getProduct(item.productId)) }}</p>
                 </ion-label>
               </ion-item>
             </div>
@@ -116,7 +116,7 @@ import Image from "@/components/Image.vue";
 import { useRouter } from 'vue-router';
 import Scanner from "@/components/Scanner.vue";
 import ImageModal from '@/components/ImageModal.vue';
-import { showToast, getProductIdentificationValue } from '@/utils'
+import { showToast, productHelpers } from '@/utils'
 import { translate } from '@/i18n'
 
 export default defineComponent({
@@ -265,7 +265,7 @@ export default defineComponent({
       checkmarkDone,
       locationOutline,
       store,
-      getProductIdentificationValue,
+      productHelpers,
       router
     };
   },
